backend/requirements.txt
fastapi
uvicorn
pymongo
google-generativeai
python-dotenv
requests
